misc fixed

+++ mk-emmc-image.sh
@@ -77,6 +77,13 @@ EOF
 download_img ${TARGET_OS}
 download_img eflasher
 
+# Automatically re-run script under sudo if not root
+if [ $(id -u) -ne 0 ]; then
+	echo "Re-running script under sudo..."
+	sudo "$0" "$@"
+	exit
+fi
+
 ./mk-sd-image.sh eflasher && \
 	./tools/fill_img_to_eflasher out/${SOC}-eflasher-$(date +%Y%m%d).img ${SOC} $@ && { 
 		rm -f out/${SOC}-eflasher-$(date +%Y%m%d).img
